tests: move mocked text results fixtures to submodule
